Keep CesiumEditor AsyncSystem around forever.
Async++ doesn't do anything to control the lifetime of its schedulers,
so it requires that the client ensure schedulers live longer than any
Future that might try to run a continuation with a scheduler. This is
really hard to ensure. After a few false starts, this commit just makes
CesiumEditor's AsyncSystem static so that it sticks around until
shutdown.